+++ test/tests-api.test.js
@@ -15,13 +15,13 @@ describe('public tests API', function() {
         it('should throw an error if first argument is not a string', function() {
             (function () {
                 this.context.suite(123, function() {});
-            }.must.throw());
+            }.must.throw(TypeError));
         });
 
         it('should throw an error if second argument is not a function', function() {
             (function() {
                 this.context.suite('name');
-            }.must.throw());
+            }.must.throw(TypeError));
         });
 
         it('should create new suite with corresponding name', function() {
@@ -65,6 +65,14 @@ describe('public tests API', function() {
         } 
 
         describe('setUrl', function() {
+            it('should throw if argument is not a string', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.setUrl({not: 'a string'});
+                    });
+                }.bind(this)).must.throw(TypeError);
+            });
+
             it('should set url property', function() {
                 this.context.suite('name', function(suite) {
                     suite.setUrl('http://example.com');
@@ -77,6 +85,22 @@ describe('public tests API', function() {
         });
 
         describe('setCaptureElements', function() {
+            it ('should throw if selector is not a string', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.setCaptureElements({everything: true});
+                    });
+                }.bind(this)).must.throw(TypeError);
+            });
+
+            it ('should throw if selector in array is not a string', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.setCaptureElements([{everything: true}, '.selector']);
+                    });
+                }.bind(this)).must.throw(TypeError);
+            });
+
             it('should set captureSelectors property', function() {
                 this.context.suite('name', function(suite) {
                     suite.setCaptureElements('.selector');
@@ -110,9 +134,41 @@ describe('public tests API', function() {
                 this.suite.children[0].beforeHook.must.be(func);
 
             });
+
+            it('should throw if hook is not a function', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.before('the dawn');
+                    }); 
+                }.bind(this)).must.throw(TypeError);
+            });
         });
 
         describe('capture', function() {
+            it('should throw if first argument is not passed', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.capture({not: 'a string'});
+                    });
+                }.bind(this)).must.throw(TypeError);
+            });
+
+            it('should throw if second argument is not a function', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.capture('state', 'make me a sandwich');
+                    });
+                }.bind(this)).must.throw(TypeError);
+            }); 
+
+            it('should not throw if second argument is absent', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.capture('state');
+                    });
+                }.bind(this)).must.not.throw();
+            });
+
             it('should create named state', function() {
                 this.context.suite('name', function(suite) {
                     suite.capture('state');
@@ -151,6 +207,46 @@ describe('public tests API', function() {
         });
 
         describe('skip', function() {
+            it('should throw if argument is not a string nor object', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.skip(123);
+                    });
+                }.bind(this)).must.throw(TypeError);
+            });
+            
+            it('should throw if argument is array with non-string or non-object', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.skip([123]);
+                    });
+                }.bind(this)).must.throw(TypeError);
+            });
+
+            it('should throw if argument is an object and browser name is not specified', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.skip({iHaveNo: 'name'});
+                    });
+                }.bind(this)).must.throw(Error);
+            });
+
+            it('should throw if browser name is not a string', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.skip({name: true});
+                    });
+                }.bind(this)).must.throw(TypeError);
+            });
+
+            it('should throw if browser version is not a string', function() {
+                (function() {
+                    this.context.suite('name', function(suite) {
+                        suite.skip({name: 'browser', version: {major: 42}});
+                    });
+                }.bind(this)).must.throw(TypeError);
+            });
+
             it('should mark suite as skipped', function() {
                 this.context.suite('name', function(suite) {
                     suite.skip();